+++ CHANGELOG.rst
@@ -4,6 +4,40 @@ Changelog
 
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
+1.4.0 (2020-12-07)
+----------------------------------------------
+
+**Added**
+
+* OSGi bundle header creation in MANIFEST.mf
+* Vaadin 8 compatibility
+* Liferay 7 compatibility
+
+**Fixed**
+
+* Calls to openBIS ingestion service are replaced with openBIS API V3 creation functionality
+
+**Dependencies**
+
+* Update: Java version from 8 to 11
+* Update: Vaadin version from 7.7.8 to 8.12.0
+* Update: Liferay version from 6.2.5 to 7.3.5
+* Update: openBIS Client Lib version from 18.06.2 to 20.10.0
+
+* Added: QBiC Liferay Settings - 1.0.0
+* Added: openBIS API Batteries Included - 20.10.0
+* Added: Liferay Portal Kernel - 9.8.0
+* Added: BND Maven Plugin in the build process - 5.2.0
+
+* Removed: Liferay Portal Service
+* Removed: Liferay Util Java
+* Removed: QBiC Portlet Parent Pom
+
+**Deprecated**
+
+* Vaadin 7 compatibility
+* Liferay 6 compatibility
+
 
 1.3.0 (2020-11-30)
 ----------------------------------------------